28.05.2015 17:43

+++ SportBlog/src/main/java/com/sport/blog/dao/GeneralDAO.java
@@ -8,7 +8,7 @@ public interface GeneralDAO<E> {
 
     public void updateElement(E element);
 
-    public E getElementByID(Long elementId);
+    public E getElementByID(Integer elementId);
 
     public List<E> getAllElements();
