+++ SportBlog/src/main/java/com/sport/blog/daoImpl/GeneralDaoImpl.java
@@ -39,7 +39,7 @@ public abstract class GeneralDaoImpl<E> implements GeneralDAO<E> {
 		}
 	}
 
-	public E getElementByID(Long elementId) {
+	public E getElementByID(Integer elementId) {
 		Session session = null;
 		E element = null;
 		try {
